define torch_dtype=16

+++ src/models/model.py
@@ -1,8 +1,9 @@
 from transformers import AutoModelForCausalLM
+import torch
 
 def get_model(**kwargs):
     if kwargs.pop("model_parallel")==True:
-        model = AutoModelForCausalLM.from_pretrained(device_map="auto", **kwargs)
+        model = AutoModelForCausalLM.from_pretrained(device_map="auto", torch_dtype=torch.float16, **kwargs)
     else:
-        model = AutoModelForCausalLM.from_pretrained(**kwargs)
+        model = AutoModelForCausalLM.from_pretrained(torch_dtype=torch.float16, **kwargs)
     return model
